Add matricies and transformations

+++ src/camera.rs
@@ -11,8 +11,8 @@ pub struct Camera {
 }
 
 impl Camera {
-    pub fn new(origin: Point, vfov: f64, aspect: f64, height: u32) -> Self {
-        let theta = vfov * std::f64::consts::PI / 180.0;
+    pub fn new(origin: Point, vfov: f32, aspect: f32, height: u32) -> Self {
+        let theta = vfov * std::f32::consts::PI / 180.0;
         let half_height = (theta / 2.0).tan();
         
         let viewport_height = 2.0 * half_height;
@@ -29,13 +29,13 @@ impl Camera {
             vertical,
             lower_left_corner: origin - horizontal / 2. - vertical / 2. - Vector::new(0., 0., focal_length),
             height,
-            width: (height as f64 * aspect) as u32,
+            width: (height as f32 * aspect) as u32,
         }
     }
 
     pub fn ray_for_pixel(self, x: u32, y: u32) -> Ray {
-        let u = (x as f64 + 0.5) / self.width as f64;
-        let v = (y as f64 + 0.5) / self.height as f64;
+        let u = (x as f32 + 0.5) / self.width as f32;
+        let v = (y as f32 + 0.5) / self.height as f32;
 
         let point_on_screen = self.lower_left_corner + u * self.horizontal + v * self.vertical;
