+++ src/intersection.rs
@@ -2,7 +2,7 @@ use crate::{intersectable::Intersectable, point::Point};
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Intersection {
-    pub t: f64,
+    pub t: f32,
     pub point: Point,
     pub object: Intersectable,
 }
